requirements.txt
appdirs==1.4.0
Django==1.10.1
django-cors-headers==1.1.0
djangorestframework==3.4.7
gunicorn==19.6.0
packaging==16.8
pyparsing==2.1.10
six==1.10.0
